// src/lib/schema.ts
export const DoctorSchema = z.object({
    name: z
        .string()
        .trim()
        .min(2, "Name must be at least 2 characters")
        .max(50, "Name must be at most 50 characters"),
    phone: z.string().min(10, "Enter phone number").max(10, "Enter phone number"),
    email: z.string().email("Invalid email address."),
    address: z
        .string()
        .min(5, "Address must be at least 5 characters")
        .max(500, "Address must be at most 500 characters"),
    specialization: z.string().min(2, "Specialization is required."),
    license_number: z.string().min(2, "License number is required"),
    type: z.enum(["FULL", "PART"], { message: "Type is required." }),
    department: z.string().min(2, "Department is required."),
    img: z.string().optional(),
    password: z
        .string()
        .min(8, { message: "Password must be at least 8 characters long!" })
        .optional()
        .or(z.literal(""))
})

export const workingDaySchema = z.object({
    day: z.enum(["monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"]),
    start_time: z.string(),
    close_time: z.string()
})
export const WorkingDaysSchema = z.array(workingDaySchema).optional()

export const StaffSchema = z.object({
    name: z
        .string()
        .trim()
        .min(2, "Name must be at least 2 characters")
        .max(50, "Name must be at most 50 characters"),
    role: z.enum(["NURSE", "LAB_TECHNICIAN"], { message: "Role is required." }),
    phone: z.string().min(10, "Contact must be 10-digits").max(10, "Contact must be 10-digits"),
    email: z.string().email("Invalid email address."),
    address: z
        .string()
        .min(5, "Address must be at least 5 characters")
        .max(500, "Address must be at most 500 characters"),
    license_number: z.string().optional(),
    department: z.string().optional(),
    img: z.string().optional(),
    password: z
        .string()
        .min(8, { message: "Password must be at least 8 characters long!" })
        .optional()
        .or(z.literal(""))
})
// ...
export const ServicesSchema = z.object({
    service_name: z.string({ message: "Service name is required" }),
    price: z.string({ message: "Service price is required" }),
    description: z.string({ message: "Service description is required" })
})